Using VMs/requirements.txt
google.cloud.storage